requirements.txt
streamlit==1.32.0
pandas==1.5.3
psycopg2-binary==2.9.9
matplotlib==3.7.5
boto3
seaborn